+++ packages/mvvm-pack/lib/amdbootstrap.module.js
@@ -31,7 +31,6 @@
     var ret = (typeof factory == 'function')
       ? factory.apply(mod, [cmlrequire, mod.exports, mod])
       : factory;
-
     if (ret) {
       mod.exports = ret;
     }